feat: add get categories endpoint
